hw3-baltazarO-master/requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
Click==7.0
coverage==4.5.4
dnspython==1.16.0
Flask==1.1.1
gunicorn==20.0.0
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mongoengine==0.18.2
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
py==1.8.0
pymongo==3.9.0
pyparsing==2.4.5
pytest==5.2.3
six==1.13.0
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
